[Docs] Added new papers for benchmark searching
After deciding that DASPfind (Ba-Alawi, 2016) wasn't going to be our
benchmark for our method, we proceded to search for other options and
found some method based in NBI (Network Bipartite Inference) that are
really interesting candidates.

Been thinking in the posibility in creating a hybrid methos using NBI
and out method based in shortest path.