[Tools] Small script to check misc cases in preds
We know about the existance of some miscellaneous cases in our
predictive method, were the shortest path between a ligand and it's
predicted target has 3 steps. This script creates a small figure
denoting the prediction pair, the shortest path and the neighbors that
this node pair share. This enables us to hand check some of this
predictions.